When dumping out a structure and determining the largest space needed for name and value take into account the size of the output window. Numbers aren't a problem, but strings can be so truncate them if necessary.

+++ src/rmcneal.com/uscsi/readcap.go
@@ -15,7 +15,7 @@ func diskinfoReadCap(d *diskInfoData) {
 		totalBlocks := converter.getInt64()
 		converter.setOffsetCount(8, 4)
 		blockSize := converter.getInt64()
-		d.capacity = totalBlocks * blockSize
+		d.capacity = uint64(totalBlocks * blockSize)
 	}
 }
